refactor the code, added part of the documentation
did quite some changes in this commit. First is importing the openapi documentation using Redoc and go-swagger, so the dependency increases. I refactored the code so that each file has fewer code, making it easier to debug. CORS support is added by using gorilla handlers, this gives a global support to better handle the prefetch packages with just one line of code.